connector/requirements.txt
# Bank-Side Connector Dependencies
# ================================

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# JWT authentication
PyJWT>=2.8.0
cryptography>=41.0.0

# Image processing
Pillow>=10.1.0

# Development/Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Optional: Rate limiting (uncomment if needed)
# slowapi>=0.1.8
